Add a header for testing git commits without -m
This time I'm testing out committing while wrapping my lines after a
certain amount of characters. Hopefully it's easier to read and this
explanation will be easier to understand when someone reads through it.

I still don't know exactly how many paragraphs are supposed to be in a
longer commit. So I'm hoping this is correct.

Anyways, this is a long commit. I'm soley doing this commit to learn.